+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v0.2.0-dev
 
+* Rename `Redix.NetworkError` to `Redix.ConnectionError` (as it's more generic
+  and more flexible).
 
 ## v0.1.0